// src/app/modules/tecnologia-informacao/controle-linhas/models/controle-linha.ts
export interface ControleLinha {
  matricula: number;
  codLinha: number;
  numLinha: number;
  descricao: string;
  codEscritorio: number;
  nomeEscritorio: string;
  codEmpresa: number;
  nomeEmpresa: string;
  codSituacao: number;
  dsSituacao: string;
  valor: number;
  numContrato: string;
  dataVencimentoContrato: string;
  codUsuario: string;
  nomeUsuario: string;
  dataCadastro: string;
  plano: string;
  gestorFun: string;
  dados: string;
  conta: string;
  codChip: string;
  numConta: number;
  senha: string;
  status: string;
  documento: string;
  qtDocumentos: number;
}